[ide] Use Dune to support separate CoqIDE build.
Packagers have long wanted to build CoqIDE against the installed Coq
libraries instead of having to rebuild parts of Coq itself.

In fact, this is indeed possible since since some time, as for example
other topleves in OPAM such as `coq-serapi` do.

We thus provide the needed `dune` files so CoqIDE can be built
separately by issuing `jbuilder build -p coqide`.

This can be backported to 8.8.1, but IMO protocol files should be
moved to its own directory as done in `master`; I don't see a lot of
risk on it tho.

Note that as of today, the resulting files have to be installed in the
same place where the `coqworker*` binaries are installed.  This could
be improved by tweaking the search heuristic of coqide / coqlib.

The Dune-generated `coqide.install` file does the right thing in that
regard.