[doc] Build ML API documentation artifact.
This is not optimal for we have to rebuild the `.cmi` as
`ocamldoc` cannot yet use the `_install_ci/` directory.

Overall the `mli` documentation is in a sorry state, however, I think
this is a first step in order to improve it.

Note that the `ml-doc` target seems broken in OCaml 4.07.0, needs
investigation.

cc: #7155
(cherry picked from commit 5cc3e117ed0264e35a46b6e3bb416b4b7f66b127)