Checking if low-level name printers are used on purpose or not.
In particular we check if really used for internal debugging purpose
or to display a message to the user. In the latter case, we replace it
(when possible) by a higher-level printer (e.g. printing foo instead
of Top.foo). In the former case, we clarify that the use is a
debugging use.

Still not perfect (see a few FIXME).

(cherry picked from commit 39a10cba3d610c6f12438084c5de7c1217c8fe94)