[coqtop] Give priority to stdlib load path over current directory
When initilazing the load path, coqtop adds implicit bindings for stdlib
and for the current directory (-R stdlib Coq -R . ""). In case of
a clash, the binding of the current directory had priority, which makes
it impossible to edit stdlib files (when they Require files from the
same directory) using PG, or from CoqIDE when launched from the directory
containing the file.

Example to reproduce the problem:

```
cd plugins/omega
coqide Omega.v
```

some of the Requires will fail.