Cleanup reduction effects: they only work on constants.
We only ever call `reduction_effect_hook` on constants, so there's no
point allowing it to be declared with globrefs. There is also no point
using a constr map instead of constant map.

(Technically there was a call of the effect hook on projections, but
that can never match a globref so it was useless)