Introduce a team of code owners for the CI system.
This means that all the members of the team will receive
a review request for PRs on the CI, but only one of them
will need to approve the PR, and this will remove the
review request for the others.

Currently the team contains Emilio and Gaetan, the two
former code owners of these files. It makes sense to start
experimenting on this component since they had already
decided to make their role symmetric.

Updating the list of maintainers can be done by updating
the list members, and without changing the CODEOWNER file.