Update/improve two aspects of the merging process.
Following discussion in #7042, we write down an advice to give time
for last comments before merging potentially controversial PRs.

We document a practice that is becoming standard in order to improve
PR merging time: some other maintainer can self-assign if the main
maintainer is not responding.

Encourage component maintainers to announce when they won't be available.

We document the practice of code owner teams.

Co-authored-by: Hugo Herbelin <Hugo.Herbelin@inria.fr>