(Partially) Revert "Make Environ.globals abstract."
This (partially) reverts commit 3984f3c1db51f7b788ad49eafb7647774e8d1f53.

This broke clients wanting to inspect the enviroment, see
https://github.com/coq/coq/pull/7745#issuecomment-411597610

There is a need for clients to inspect the global environment, we keep
the record private as to concerns regarding its use, so even if no
function in the kernel is taking a `globals` as an input, we transmit
to clients its read-only nature.

We take the opportunity to refactor the record into a module with
scoped constructors.