Modules and definitions now share the same namespace.
This prevents naming a definition and a module the same. This will turn out to
be useful when we implement side-effects as local definitions, so that the
nametab can reuse the same syntax for module and definition qualification.

Furthermore the current situation is a bit clunky when it comes to named
universes, as both a definition and a module can generate qualified universe
names. This can result in strange situations.

The only places that relied on the old behaviour were tests, either to
precisely check this property, or that use standard naming schemes that
clashed with one another (foo, bar, etc.).