Update minimum required dependency versions of Sphinx doc.
The minimum required versions of the Sphinx-related (and ANTLR)
Python packages for Coq 8.10 were chosen as the lower bound
between what is currently in Debian Buster and in NixOS 18.09
Jellyfish (in practice the lower bound was always met by NixOS
18.09 Jellyfish).

These minimum required versions were documented.

In the docker image used by GitLab CI, we install these
Python packages through pip as this allows us to pin them
to these specific versions.

In Travis, we let them unspecified to always test the
latest versions.

Finally, we also add the new dependencies of the Sphinx
PDF manual.