Fix Numeral Notations (2/4 - exceptions, usr_err)
Switch to using exceptions rather than user errors.  This will be
required because the machinery for printing constrs is not available in
notation.ml, so we move the error message printing to himsg.ml instead.

This is commit 2/4 in the fix for #8401.

(cherry picked from commit 4888eb7ea720eaa34af8df2769fa300f1ea37173)