Fix Numeral Notations (3/4 - moving more stuff)
Move most of the rest of the stuff from numeral.ml to notation.ml.  Now
that we use exceptions to print error messages, all of the
interpretation code for numeral notations can be moved to notation.ml.

This is commit 1/4 in the fix for #8401.

This is a pure cut/paste commit, modulo fixing name qualifications due
to moved things, and exposing some stuff in notation.mli (to be removed
in the next commit, where we finish the numeral notation reworking).