Define polymorphism attribute through the extensible system
Note that [attributes_of_flags] used to preserve the value of the
polymorphic field in the atts it got as argument. This probably broke
the classifier when using a polymorphism attribute since the
classifier did not look at the option returned by attributes_of_flags.

Vernacentries meanwhile overrode it in some complicated way so that
interp did get the right polymorphism. I guess we just don't test this
bit of the classifier.