[ci] [docker] Move to OPAM 2.0
The OPAM 1.2 repository has been frozen, thus we must use OPAM 2.0 if
we want to get newer versions of packages / compilers.

For now we must perform a manual install of OPAM as no packages for
Ubuntu 18.04 exist.

Note that this means nothing about the installability of Coq itself,
whose OPAM repository should remain in 1.2 format for quite a long
period.