Miscellaneous refinements/cleaning of module printing.
This refines the fix to #2169 by distinguishing the short and
non-short printing modes.

This repares functionalization of printers by always passing env
rather than setting env to None in short mode. This is not strictly
necessary for the env which is not used for printing global references
but it shall be more consistent in style when passing e.g. the nametab
functionally.

This also removes the fallback printer used in case of error while
printing; otherwise said, we make the assumption that the non-short
mode of the module printer is not buggy.