[dune] Support compiling libraries with the native compiler.
This is mostly a proof of concept/RFC; we likely want to refine this
more, so not ready for merge yet.

In particular it has the problem of dropping the `.coq_native` prefix,
which is not such a big deal for hygienic builds but it is cumbersome
for the rest.

This is due to a limitation of the build language, such limitation
doesn't exist on the native Coq Dune mode.

Ideally we would like to isolate native runtime dependencies on its
own library. That's not too hard to do with Dune's `copy_files`.

Another thing to do in the native Coq mode is to add a `(mode native)`
flag.