[dune] Refactor files following advice from upstream.
After some discussion upstream, I think that for the current a
multi-package setup in a single repository this setup saves a bit
work.

The most problematic bit is the `-rectypes` flag; its status is not
saved per-library so we must either duplicate the flags in the coqide
scope (scope == dune-project file), or unify its scope with Coq. We do
this last option as this seems like the easiest thing for now.

Changes:

- Move coqide.opam file to the root, use a single scope/`dune-project` file.
- Remove `dune-workspace`, this should be owned by the developer / user.