[printing] Collect Printing Options in a record (c.f. #6560)
This PR was originally written by Paul Steckler (#6698), and updated,
rebased, and modified by me and others.

The main changes are the placement of all printing options in a
centralized record, thus the printing state is accessible in a more
convenient place. See CHANGES for printing semantic changes.

The most important change is that `Set Printing All` is not an ad-hoc
operation anymore, but just setting all options to "display".

There remain some important TODOs, apart from code review:

- Update the reference manual, diff in `pending_refman_update.diff`.
- Check IDE changes by an expert.

Co-authored-by: Paul Steckler <steck@stecksoft.com>