[printing] Handle groups of printing options in a more standard way.
We remove the special vernaculars and locality handling code in favor
of a standard treatment for `Printing All` and friends.

We use a single saved register thus unsetting may play some tricks on the
user if they are looking for it, on the other hand that makes it
easier to understand.