[coq_makefile] Fail when no `CONF_FILE` is present.
When a `Makefile.conf` file is not present, `coq_makefile` used to try
to re-generate it. This results in #8731.

Instead, we require that the `.conf` file is present, or else we fail;
this means that users cannot distribute generated makefiles
anymore. Indeed, this makes sense as `coq_makefile` writes hardcoded
paths into makefiles.