Enable fragile pattern warning in cclosure
This file is already mostly in the required style so I wanted to see
what it looks like.

For a couple matches I locally disabled the warning as it was too
annoying otherwise (`when` clauses are especially annoying).

There are a couple places where I think it clearly looks better (eg
assoc_defined at the beginning of the file) but overall I'm not all
that convinced.
What do other people think?