[core] Remove custom backtrace module in favor of OCaml's.
We take advantage some newer primitives such as `reraise`.  This is
mainly an RFC and will need fine tuning, but it seems to provide
better backtraces and I do prefer the cleanup of the current API.

There is a catch, `reraise` won't work properly if another exception
is called in the exception handler itself, in this case we need to use
some extra protection.

I have tried to preserve places that mess with the `exninfo`, however
all that stuff is suspicious and I'm afraid that hack that could stop
being supported upstream at any time. IMHO we should migrate from it.

In particular, it is beneficial that `CErrors` doesn't not depend now
on `Exninfo`, as this used to give problems in the past.

Should fix some day #6446.